[Infra] Add shared OtlpHttpCollector class
Add a shared `OtlpHttpCollector` class to re-use between the Android, Blazor and iOS tests.
