+++ test/OpenTelemetry.Apple.Tests/AppleEndToEndTests.cs
@@ -3,6 +3,7 @@
 
 using OpenTelemetry.Proto.Common.V1;
 using OpenTelemetry.Proto.Logs.V1;
+using OpenTelemetry.Tests;
 
 namespace OpenTelemetry.Apple.Tests;
