+++ test/OpenTelemetry.BlazorWasm.Tests/BlazorWasmEndToEndTests.cs
@@ -4,6 +4,7 @@
 using Microsoft.Playwright;
 using OpenTelemetry.Proto.Common.V1;
 using OpenTelemetry.Proto.Logs.V1;
+using OpenTelemetry.Tests;
 
 namespace OpenTelemetry.BlazorWasm.Tests;
