+++ test/OpenTelemetry.BlazorWasm.Tests/OpenTelemetry.BlazorWasm.Tests.csproj
@@ -16,6 +16,7 @@
 
   <ItemGroup>
     <Compile Include="$(RepoRoot)\test\OpenTelemetry.Tests\Shared\TcpPortProvider.cs" Link="Includes\TcpPortProvider.cs" />
+    <Compile Include="$(RepoRoot)\test\Shared\OtlpHttpCollector.cs" Link="Includes\OtlpHttpCollector.cs" />
   </ItemGroup>
 
   <ItemGroup>
